+++ src/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as usePagination } from "./usePagination";
